requirements.txt
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.17.0
azure-storage-file-datalake==12.12.0
azure-storage-file-share==12.13.0
azure-storage-queue==12.6.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.9.0
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
plotly==5.15.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
PyPDF2==3.0.1
pypdfium2==4.18.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.0
rarfile==4.0
reportlab==4.0.4
requests==2.31.0
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.25.0
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
wcwidth==0.2.6
xgboost==1.7.6
XlsxWriter==3.1.2
zipp==3.15.0
